Fix removal of DigitalObjects, which contain no identifiers

+++ dc-cudami-server/dc-cudami-server-backend-jdbi/src/main/java/de/digitalcollections/cudami/server/backend/impl/jdbi/identifiable/IdentifierRepositoryImpl.java
@@ -43,6 +43,10 @@ public class IdentifierRepositoryImpl extends JdbiRepositoryImpl implements Iden
 
   @Override
   public void delete(List<UUID> uuids) {
+    if (uuids == null || uuids.isEmpty()) {
+      return;
+    }
+
     dbi.withHandle(
         h ->
             h.createUpdate("DELETE FROM " + tableName + " WHERE uuid in (<uuids>)")
